Add files via upload
Uv's are the basis for how to map textures on to a model named for the U and V axis in the editor
planar mapping is used to map an object based on a view plane either from camera, x, y, or z axis
Cylindrical mapping is useful for mapping cylinder primitives or objects closer in shape to the primitive. you do need to set the edge to close agains the other side
spherical maps are open on three sides the top and bottom as well as the edge between them you do need to close those by manipulating the handles
automatic mapping isn't always the best mode to use to map an object it can create a very jumbled mess. it's easier to use on simple objects
unfold is used to flatten the image plane to get more consistent layouts for texture editing and for mapping the uvs it makes it easier to view the faces in your object and to get it into the flat plane
some of the issues come with distortion of the textures and these are usually fixed by regrouping cutting flattening and re organizing the struc